Refactored "HTML Attribute" extensions
