+++ src/Umbraco.Web.UI.Client/src/external/tiptap/index.ts
@@ -133,7 +133,6 @@ export { StarterKit } from '@tiptap/starter-kit';
 // CUSTOM EXTENSIONS
 export * from './extensions/tiptap-figcaption.extension.js';
 export * from './extensions/tiptap-figure.extension.js';
-export * from './extensions/tiptap-html-global-attributes.extension.js';
 export * from './extensions/tiptap-text-direction-extension.js';
 export * from './extensions/tiptap-text-indent-extension.js';
 export * from './extensions/tiptap-trailing-node.extension.js';
