+++ src/Umbraco.Web.UI.Client/src/packages/tiptap/extensions/html-attr-class/index.ts
@@ -1,0 +1,1 @@
+export * from './html-attr-class.tiptap-extension.js';
