+++ src/Umbraco.Web.UI.Client/src/packages/tiptap/extensions/html-attr-dataset/index.ts
@@ -1,0 +1,1 @@
+export * from './html-attr-dataset.tiptap-extension.js';
