+++ src/Umbraco.Web.UI.Client/src/packages/tiptap/extensions/html-attr-id/index.ts
@@ -1,0 +1,1 @@
+export * from './html-attr-id.tiptap-extension.js';
