+++ src/Umbraco.Web.UI.Client/src/packages/tiptap/extensions/html-attr-style/index.ts
@@ -1,0 +1,1 @@
+export * from './html-attr-style.tiptap-extension.js';
