+++ src/Umbraco.Web.UI.Client/src/packages/tiptap/extensions/index.ts
@@ -3,6 +3,10 @@ export * from './tiptap-toolbar-element-api-base.js';
 export * from './anchor/index.js';
 export * from './bubble-menu/index.js';
 export * from './character-map/index.js';
+export * from './html-attr-class/index.js';
+export * from './html-attr-dataset/index.js';
+export * from './html-attr-id/index.js';
+export * from './html-attr-style/index.js';
 export * from './html-tag-div/index.js';
 export * from './html-tag-span/index.js';
 export * from './table/index.js';
